feat(db): add prisma and supabase adapters
