+++ package.json
@@ -34,6 +34,7 @@
     "@prisma/extension-accelerate": "^2.0.1",
     "@react-aria/live-announcer": "3.4.3",
     "@react-spring/web": "^10.0.1",
+    "@supabase/supabase-js": "^2.53.0",
     "@svgr/webpack": "^8.1.0",
     "@tanstack/react-query": "^5.81.5",
     "@use-gesture/react": "^10.3.1",
